25/03

+++ backend/src/todos/use-cases/delete-todo.use-case.ts
@@ -1,10 +1,11 @@
-import { Injectable, Logger } from "@nestjs/common";
-import { DeleteTodoRepository } from "../repository";
+import { Injectable, Logger, NotFoundException } from "@nestjs/common";
+import { DeleteTodoRepository, findtodobyidrepository } from "../repository";
 
 @Injectable()
 export class DeleteTodoUseCase {
     constructor(
       private readonly deleteTodoRepository: DeleteTodoRepository,
+      private readonly findtodobyidrepository: findtodobyidrepository,
       private readonly logger: Logger,
     ) {}
 
@@ -12,7 +13,12 @@ export class DeleteTodoUseCase {
     async execute( id: string) {
         try {
             this.logger.log('Deleting toDo...');
-            const todo = await this.deleteTodoRepository.Delete(id);
+            const todo = await this.findtodobyidrepository.findbyid(id);
+            if (!todo) {
+                throw new NotFoundException('ToDo not found')
+            }
+
+            await this.deleteTodoRepository.Delete(id);
             this.logger.log('ToDo deleted sucessfully');
             return todo;
         } catch (error) {
